requirements.txt
nanobot-ai>=0.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4
pyyaml>=6.0
requests>=2.31
pyautogui>=0.9.54
playwright>=1.48.0
python-dotenv>=1.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
pandas>=2.1.0
